multi_agentes_paralelo/requirements.txt
langgraph
google-generativeai
python-dotenv
langchain-google-genai
langchain